Update psk.lua
Went through the API for functionality that deals with guild rosters.  The idea now is to display a list of guild members (one that is persistent, not wiped between logins), and display a class icon next to their name, with a hover tooltip.  

I'll test it tonight when I get home.