Update psk.lua
Because classic era doesn't have the same returns to GetGuildRosterInfo that retail does, I'm trying to find a workaround to getting coords for class textures.  I'll run this tonight (unless you beat me to it).  

Additionally, the roster ONLY records people at max level in the database.  They will show as gray (offline), online and not in raid (yellow), or online and in raid (green).

Finally, I added a cheeky bit to spit out a bid line when someone bids in raid (or guild chat, for testing purposes).