Update psk.lua
Sorted guild roster by who's online first (I know it doesn't work with KSK, but I have an idea about that).  Added a refresh button to manually refresh the roster and got rid of the auto-refresh because it was spamming the console every time someone logged in/out/joined/gquit (especially last night).  I also added a little razzledazzle to the button and gave it the AH window sound and made it flash when you click it...lol.